pyproject.toml
[project]
name = "python-kel102"
version = "0.1.0"
description = ""
authors = [
    {name = "Philipp Bachmann",email = "philippbachmann08@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
packages = [{include = "python_kel102", from = "src"}]
dependencies = [
    "pyserial (>=3.5,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
